fix: fixing schemas filter and removing react rule
