+++ src/components/SearchableList/Container.tsx
@@ -29,7 +29,6 @@ function SearchableListContainer({
 	useEffect(() => {
 		const processedList = getList()
 		setList(processedList)
-		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [searchTerm, orderAsc, paginationCurrentPage, initialList])
 
 	useEffect(() => {
@@ -40,7 +39,6 @@ function SearchableListContainer({
 		} else {
 			setTotalPages(Math.ceil(initialList.length / paginationSize))
 		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [searchTerm, initialList])
 
 	function toggleOrder() {
